Connect to the dependencies before continuing.
